change `shape` to expect `seq` if arraymancer is installed
In case a module imports both seqmath and arraymancer and the `shape`
proc is called somewhere, there is ambiguity, because both
arraymancer.shape and seqmath.shape match for openArrays. By expecting
a `seq` specifically, we can hack our way around it without having to
change the name of the proc. Not nice though.